solidity contract changes

+++ contracts/Number.sol
@@ -4,9 +4,9 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Num{
  
- uint256 public Get_number;
+ uint  public Get_number;
 
- function SetNumber ( uint256 _Taken_number) public {
+ function SetNumber ( uint _Taken_number) public {
      Get_number = _Taken_number;
  }
